+++ src/__init__.py
@@ -2,6 +2,6 @@ __version__ = "0.6"
 
 
 def main():
-    from lice.core import main
+    from src.core import main
 
     main()
