+++ src/template-agpl3.txt
