+++ src/template-bsd3.txt
