+++ src/template-cc0.txt
