+++ src/template-cc_by_nc-header.txt
