+++ src/template-cc_by_nc_nd.txt
