+++ src/template-cc_by_nd-header.txt
