+++ src/template-lgpl.txt
