+++ src/template-wtfpl-header.txt
