+++ src/template-wtfpl.txt
